Added Drill Defender, and began creating resource system and displays.
The Drill Defender has been added, the Drill will create Resources which will be used to build other Defenders. The beginning of the implementation of the Resource system is in, but has no functionality other than a display that doesn't change currently. The grid height has been reduced to 5 from 6, and I plan on updating the background sprite to be more in line with the change and overall making the background look better.