Winning, Losing, and a level progress slider
Levels within the game now have a set length, and once that length has been reached and all Attackers are destroyed, the player can win. On the contrary, if the player runs out of lives, the player will lose. There is a Win canvas with graphics for winning, and a Lose canvas for losing. Once the player has won, the next level will load.