Options menu, music and difficulty
An Options menu has been implemented, allowing the player to set the music volume and difficulty of the game. There is now music for the entire game, and for winning a level. A difficulty parameter has now been added, slightly altering the prices of Defenders to increase or decrease the challenge of the game.